[chore/#19] chore 패키지 변경
