+++ app/src/main/java/sopt/org/millie/presentation/setting/navigation/SettingNavigator.kt
@@ -1,4 +1,4 @@
-package sopt.org.millie.presentation.shelf
+package sopt.org.millie.presentation.setting.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,17 +6,18 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import sopt.org.millie.core.navigation.MainTabRoute
+import sopt.org.millie.presentation.setting.SettingScreenRoute
 
-fun NavController.navigateToShelf(
+fun NavController.navigateToSetting(
     navOptions: NavOptions,
 ) {
-    navigate(MainTabRoute.Shelf, navOptions)
+    navigate(MainTabRoute.Setting, navOptions)
 }
 
-fun NavGraphBuilder.shelfNavGraph(
+fun NavGraphBuilder.settingNavGraph(
     paddingValues: PaddingValues,
 ) {
-    composable<MainTabRoute.Shelf> {
-        ShelfScreenRoute(paddingValues)
+    composable<MainTabRoute.Setting> {
+        SettingScreenRoute(paddingValues)
     }
 }
