+++ app/src/main/java/sopt/org/millie/presentation/shelf/navigation/ShelfNavigator.kt
@@ -1,4 +1,4 @@
-package sopt.org.millie.presentation.wandok
+package sopt.org.millie.presentation.shelf.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,17 +6,18 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import sopt.org.millie.core.navigation.MainTabRoute
+import sopt.org.millie.presentation.shelf.ShelfScreenRoute
 
-fun NavController.navigateToWandok(
+fun NavController.navigateToShelf(
     navOptions: NavOptions,
 ) {
-    navigate(MainTabRoute.Wandok, navOptions)
+    navigate(MainTabRoute.Shelf, navOptions)
 }
 
-fun NavGraphBuilder.wandokNavGraph(
+fun NavGraphBuilder.shelfNavGraph(
     paddingValues: PaddingValues,
 ) {
-    composable<MainTabRoute.Wandok> {
-        WandokScreenRoute(paddingValues)
+    composable<MainTabRoute.Shelf> {
+        ShelfScreenRoute(paddingValues)
     }
 }
