+++ app/src/main/java/sopt/org/millie/presentation/today/navigation/TodayNavigator.kt
@@ -1,4 +1,4 @@
-package sopt.org.millie.presentation.setting
+package sopt.org.millie.presentation.today.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,17 +6,18 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import sopt.org.millie.core.navigation.MainTabRoute
+import sopt.org.millie.presentation.today.TodayScreenRoute
 
-fun NavController.navigateToSetting(
+fun NavController.navigateToToday(
     navOptions: NavOptions,
 ) {
-    navigate(MainTabRoute.Setting, navOptions)
+    navigate(MainTabRoute.Today, navOptions)
 }
 
-fun NavGraphBuilder.settingNavGraph(
+fun NavGraphBuilder.todayNavGraph(
     paddingValues: PaddingValues,
 ) {
-    composable<MainTabRoute.Setting> {
-        SettingScreenRoute(paddingValues)
+    composable<MainTabRoute.Today> {
+        TodayScreenRoute(paddingValues)
     }
 }
