+++ app/src/main/java/sopt/org/millie/presentation/wandok/navigation/WandokNavigator.kt
@@ -1,4 +1,4 @@
-package sopt.org.millie.presentation.today
+package sopt.org.millie.presentation.wandok.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,17 +6,18 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import sopt.org.millie.core.navigation.MainTabRoute
+import sopt.org.millie.presentation.wandok.WandokScreenRoute
 
-fun NavController.navigateToToday(
+fun NavController.navigateToWandok(
     navOptions: NavOptions,
 ) {
-    navigate(MainTabRoute.Today, navOptions)
+    navigate(MainTabRoute.Wandok, navOptions)
 }
 
-fun NavGraphBuilder.todayNavGraph(
+fun NavGraphBuilder.wandokNavGraph(
     paddingValues: PaddingValues,
 ) {
-    composable<MainTabRoute.Today> {
-        TodayScreenRoute(paddingValues)
+    composable<MainTabRoute.Wandok> {
+        WandokScreenRoute(paddingValues)
     }
 }
